biblio abstract on hover
